# docs/swagger.yaml
definitions:
...
  api.DiskRemoveReservedRsp:
    properties:
      diskName:
        description: DiskName
        type: string
      removeReservedRsp:
        description: RemoveReservedRsp
        type: string
    type: object
  api.DiskRemoveReservedRspBody:
    properties:
      data:
        $ref: '#/definitions/api.DiskRemoveReservedRsp'
    type: object
  api.DiskReservedRsp:
    properties:
      diskName:
        description: DiskName
        type: string
      reservedRsp:
        description: ReservedRsp
        type: string
    type: object
  api.DiskReservedRspBody:
    properties:
      data:
        $ref: '#/definitions/api.DiskReservedRsp'
    type: object
...
  api.LocalDisk:
    properties:
      availableCapacityBytes:
        description: AvailableCapacityBytes 可用容量
        type: integer
      devPath:
        description: e.g. /dev/sdb 磁盘路径
        type: string
      hasRaid:
        description: HasRAID 是否Raid
        type: boolean
      localStoragePooLName:
        description: LocalStoragePooLName 存储池名称
        type: string
      reserved:
        description: Reserved 是否预留
        type: boolean
      state:
        description: 'Possible state: Claimed, UnClaimed, Inuse, Released, Reserved
          状态'
        type: string
      totalCapacityBytes:
        description: TotalCapacityBytes 总容量
        type: integer
      type:
        description: 'Supported: HDD, SSD, NVMe, RAM 磁盘类型'
        type: string
    type: object
  api.LocalDiskListByNode:
    properties:
      items:
        description: localDisks 节点磁盘列表
        items:
          $ref: '#/definitions/api.LocalDisk'
        type: array
      nodeName:
        description: nodeName 节点名称
        type: string
      pagination:
        $ref: '#/definitions/api.Pagination'
        description: |-
          // localDisks 节点磁盘列表
          LocalDisksItemsList LocalDisksItemsList `json:"items,omitempty"`
          page 信息
    type: object
...
  api.NodeDiskListByPool:
    properties:
      items:
        description: localDisks 节点磁盘列表
        items:
          $ref: '#/definitions/api.LocalDisk'
        type: array
      nodeName:
        description: nodeName 节点名称
        type: string
      page:
        $ref: '#/definitions/api.Pagination'
        description: page 信息
      storagePoolName:
        description: StoragePoolName 存储池名称
        type: string
    type: object
  api.NodeRemoveReserveReqBody:
    properties:
      diskName:
        type: string
      nodeName:
        type: string
    type: object
  api.NodeReserveReqBody:
    properties:
      diskName:
        type: string
      nodeName:
        type: string
    type: object
...
  api.Pagination:
    properties:
      page:
        description: 当前页索引，从 1 开始，为 0 时，会自动重置为默认值 constants.DefaultPage
        type: integer
      pageSize:
        description: |-
          每页数据量，为 -1 时表示查询全部，为 0 时会重置为默认值
          constants.DefaultPageSize
        type: integer
      pages:
        description: 总页数
        type: integer
      total:
        description: 总共有多少条目，请求时可以不用传递
        type: integer
    type: object
  api.RspFailBody:
    properties:
      description:
        type: string
      errcode:
        type: integer
    type: object
  api.StorageNode:
    properties:
      allocatedHDDCapacityBytes:
        description: allocatedHDDCapacityBytes HDD已经分配存储量
        type: integer
      allocatedSSDCapacityBytes:
        description: allocatedSSDCapacityBytes SSD已经分配存储量
        type: integer
      driverStatus:
        description: driver status 驱动状态  运行中（Ready）,维护中（Maintain）, 离线（Offline）
        type: string
      ip:
        description: ip 节点IP
        type: string
      isRaid:
        description: IsRAID 是否Raid
        type: boolean
      name:
        description: name 节点名字
        type: string
      nodeState:
        description: node state 节点状态 运行中（Ready）,未就绪（NotReady）,未知（Unknown）
        type: string
      totalDiskCount:
        description: totalDiskCount 总磁盘数
        type: integer
      totalHDDCapacityBytes:
        description: |-
          // freeCapacityBytes LSN可分配存储容量
          FreeCapacityBytes int64 `json:"freeCapacityBytes,omitempty"`
          totalHDDCapacityBytes HDD存储总容量
        type: integer
      totalSSDCapacityBytes:
        description: totalSSDCapacityBytes SSD存储总容量
        type: integer
      usedDiskCount:
        description: usedDiskCount 已绑定磁盘数
        type: integer
    type: object
  api.StorageNodeList:
    properties:
      items:
        description: StorageNodes
        items:
          $ref: '#/definitions/api.StorageNode'
        type: array
      pagination:
        $ref: '#/definitions/api.Pagination'
        description: |-
          StorageNodesItemsList
          StorageNodesItemsList StorageNodesItemsList `json:"items,omitempty"`
          page 信息
    type: object
  api.StorageNodeListByPool:
    properties:
      items:
        description: StorageNodes
        items:
          $ref: '#/definitions/api.StorageNode'
        type: array
      page:
        $ref: '#/definitions/api.Pagination'
        description: page 信息
      storagePoolName:
        description: StoragePoolName 存储池名称
        type: string
    type: object
  api.StoragePool:
    properties:
      allocatedCapacityBytes:
        description: AllocatedCapacityBytes 存储池已经分配存储容量
        type: integer
      class:
        description: 'Supported class: HDD, SSD, NVMe 磁盘类型'
        type: string
      createTime:
        description: createTime 创建时间
        type: string
      name:
        description: 'Supported pool name: HDD_POOL, SSD_POOL, NVMe_POOL 存储池名称'
        type: string
      nodesNum:
        description: NodesNum 节点数
        type: integer
      totalCapacityBytes:
        description: TotalCapacityBytes 存储池对应存储总容量
        type: integer
    type: object
  api.StoragePoolList:
    properties:
      items:
        description: storagePools
        items:
          $ref: '#/definitions/api.StoragePool'
        type: array
      page:
        $ref: '#/definitions/api.Pagination'
        description: page 信息
    type: object
...
  api.VolumeMigrateOperation:
    properties:
      endTime:
        description: EndTime 迁移结束时间
        type: string
      name:
        description: VolumeMigrateName 迁移CRD名称
        type: string
      sourceNode:
        description: SourceNode 迁移源节点
        type: string
      startTime:
        description: StartTime 迁移开始时间
        type: string
      state:
        description: State 迁移状态
        type: string
      targetNode:
        description: TargetNode 迁移目的节点
        type: string
      volumeName:
        description: VolumeName 迁移卷名称
        type: string
    type: object
...
  api.VolumeOperationListByNode:
    properties:
      items:
        description: VolumeOperations
        items:
          $ref: '#/definitions/api.VolumeMigrateOperation'
        type: array
      nodeName:
        description: node name
        type: string
      pagination:
        $ref: '#/definitions/api.Pagination'
        description: |-
          // VolumeMigrateOperationItemsList
          VolumeMigrateOperationItemsList VolumeMigrateOperationItemsList `json:"items,omitempty"`
          page 信息
    type: object
...
  api.YamlData:
    properties:
      data:
        description: yaml data
        type: string
    type: object
info:
  contact: {}
paths:
...
  /nodes/storagenode/{nodeName}/disks:
    get:
      consumes:
      - application/json
      description: list StorageNodeDisks
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - description: state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LocalDiskListByNode'
      summary: 摘要 获取指定存储节点磁盘列表
      tags:
      - Node
  /nodes/storagenode/{nodeName}/disks/{diskName}/removereserve:
    post:
      consumes:
      - application/json
      description: post RemoveReserveStorageNodeDisk
      parameters:
      - description: reqBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.NodeRemoveReserveReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/api.DiskRemoveReservedRspBody'
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 解除磁盘预留
      tags:
      - Node
  /nodes/storagenode/{nodeName}/disks/{diskName}/reserve:
    post:
      consumes:
      - application/json
      description: post ReserveStorageNodeDisk diskname i.g sdb sdc ...
      parameters:
      - description: reqBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.NodeReserveReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/api.DiskReservedRspBody'
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 预留磁盘
      tags:
      - Node
  /nodes/storagenode/{nodeName}/migrates:
    get:
      consumes:
      - application/json
      description: get StorageNodeMigrate
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VolumeOperationListByNode'
      summary: 摘要 获取指定节点数据卷迁移任务列表
      tags:
      - Node
  /nodes/storagenodeoperations/{migrateOperationName}/yaml:
    get:
      consumes:
      - application/json
      description: get StorageNodeVolumeOperationYamlGet
      parameters:
      - description: migrateOperationName
        in: path
        name: migrateOperationName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/api.YamlData'
      summary: 摘要 获取节点数据卷操作记录yaml信息
      tags:
      - Node
  /nodes/storagenodes:
    get:
      consumes:
      - application/json
      description: list StorageNode  驱动状态 [运行中（Ready）,维护中（Maintain）, 离线（Offline)]
        , 节点状态 [运行中（Ready）,未就绪（NotReady）,未知（Unknown)]
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - description: nodeState
        in: query
        name: nodeState
        type: string
      - description: driverState
        in: query
        name: driverState
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StorageNodeList'
      summary: 摘要 获取存储节点列表
      tags:
      - Node
  /nodes/storagenodes/{name}:
    get:
      consumes:
      - application/json
      description: get StorageNode 驱动状态 [运行中（Ready）,维护中（Maintain）, 离线（Offline)] ,
        节点状态 [运行中（Ready）,未就绪（NotReady）,未知（Unknown)]
      parameters:
      - description: name
        in: path
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StorageNode'
      summary: 摘要 获取指定存储节点
      tags:
      - Node
  /pools/storagepool/{storagePoolName}/nodes:
    get:
      consumes:
      - application/json
      description: get StorageNodesGetByPoolName
      parameters:
      - description: storagePoolName
        in: path
        name: storagePoolName
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - description: nodeName
        in: query
        name: nodeName
        type: string
      - description: state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StorageNodeListByPool'
      summary: 摘要 获取指定存储池存储节点列表信息
      tags:
      - Pool
  /pools/storagepool/{storagePoolName}/nodes/{nodeName}/disks:
    get:
      consumes:
      - application/json
      description: get StorageNodeDisksGetByPoolName
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: storagePoolName
        in: path
        name: storagePoolName
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.NodeDiskListByPool'
      summary: 摘要 获取指定存储池指定存储节点磁盘列表信息
      tags:
      - Pool
  /pools/storagepools:
    get:
      consumes:
      - application/json
      description: list StoragePools
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StoragePoolList'
      summary: 摘要 获取存储池列表信息
      tags:
      - Pool
  /pools/storagepools/{name}:
    get:
      consumes:
      - application/json
      description: get Pool
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StoragePool'
      summary: 摘要 获取指定存储池基本信息
      tags:
      - Pool
...
swagger: "2.0"